+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 2.3.0
+
+- Emit event when ipl-button is right-clicked
+- Add `ipl-small-toggle` component
+
 # 2.2.1
 
 - Incredibly minor: Remove default margin from ipl-button